Add a passing test demonstrating syntax awkwardness around zero arguments.
One of the most common usages is doing the weakify/strongify dance on `self`while using a closure that takes no arguments. The test added in this commit shows how we must shoehorn this usage, which is one context with no arguments, into the closest existing Strongify method which takes one context and one argument.

When I first used Strongify, I ran into a compilation error `Cannot invoke 'strongify' with an argument list of type '(weak: <any type>, (<any type>?) -> ())'. I stumbled into the resolution seen in the passing test, which is passing an additional underscore that doesn't seem to correspond to anything in particular. I scratched my head for some time then moved on. Much later, I realize the need for an argument, ignored or not, is because there are no zero-argument method variants. The argument supports compiler inference of the `Argument1` type. The argument can be ignored, and the compiler either infers a Void type or understands enough to also ignore that type.

This is awkward and confusing, and given the frequency of this usage, it's often the first thing someone will encounter when checking out the framework. I think a significant number of people would decide to not use the framework because of it.